Updates to rw::geometry as specified in review request #108 ConvexHull3D: Change the return type of toTriMesh from ordinary to smart pointer to be consistent with what we usually to in Robwork.
OBBFactory: Updated implementation to reflect the change above.

PlainTriMesh: Added a method for adding all triangles in another PlainTriMesh.

PlainTriMesh: Added a clear method for emptying the content

Pyramid: Added explicit casts to avoid warnings on double to float conversion.

QHull3D: Updated implementation to match the new interface on ConvexHull3D

QHull3D: Added methods for getting the plane normals and offsets

Triangulate: Added method for triangulating a 3D planar polygon, instead of only a 2D and added documentation.

git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@5698 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/Pyramid.cpp
@@ -40,30 +40,30 @@ using namespace rw::math;
 TriMesh::Ptr Pyramid::createMesh(int resolution) const{
     rw::geometry::PlainTriMeshF::Ptr mesh = ownedPtr( new rw::geometry::PlainTriMeshF( 6 ) );
     // the bottom two triangles
-    (*mesh)[0].getVertex(0) = Vector3D<float>(-_widthX/2,-_widthY/2,0);
-    (*mesh)[0].getVertex(1) = Vector3D<float>(-_widthX/2, _widthY/2,0);
-    (*mesh)[0].getVertex(2) = Vector3D<float>( _widthX/2, _widthY/2,0);
+    (*mesh)[0].getVertex(0) = Vector3D<float>(static_cast<float>(-_widthX/2.0),static_cast<float>(-_widthY/2), 0);
+    (*mesh)[0].getVertex(1) = Vector3D<float>(static_cast<float>(-_widthX/2.0), static_cast<float>((float)_widthY/2.0), 0);
+    (*mesh)[0].getVertex(2) = Vector3D<float>(static_cast<float>(_widthX/2.0), static_cast<float>(_widthY/2), 0);
 
-    (*mesh)[1].getVertex(0) = Vector3D<float>( _widthX/2, _widthY/2,0);
-    (*mesh)[1].getVertex(1) = Vector3D<float>( _widthX/2,-_widthY/2,0);
-    (*mesh)[1].getVertex(2) = Vector3D<float>(-_widthX/2,-_widthY/2,0);
+    (*mesh)[1].getVertex(0) = Vector3D<float>(static_cast<float>(_widthX/2.0),static_cast<float>( _widthY/2), 0);
+    (*mesh)[1].getVertex(1) = Vector3D<float>(static_cast<float>(_widthX/2.0),static_cast<float>(-_widthY/2), 0);
+    (*mesh)[1].getVertex(2) = Vector3D<float>(static_cast<float>(-_widthX/2.0),static_cast<float>(-_widthY/2.0), 0);
 
     // the four sides
-    (*mesh)[2].getVertex(0) = Vector3D<float>( 0, 0, _height);
-    (*mesh)[2].getVertex(1) = Vector3D<float>( _widthX/2,-_widthY/2,0);
-    (*mesh)[2].getVertex(2) = Vector3D<float>( _widthX/2, _widthY/2,0);
+    (*mesh)[2].getVertex(0) = Vector3D<float>( 0, 0, static_cast<float>(_height));
+    (*mesh)[2].getVertex(1) = Vector3D<float>(static_cast<float>(_widthX/2.0),static_cast<float>(-_widthY/2.0), 0);
+    (*mesh)[2].getVertex(2) = Vector3D<float>(static_cast<float>(_widthX/2.0),static_cast<float>(_widthY/2.0), 0);
 
-    (*mesh)[3].getVertex(0) = Vector3D<float>( 0, 0, _height);
-    (*mesh)[3].getVertex(1) = Vector3D<float>( _widthX/2, _widthY/2,0);
-    (*mesh)[3].getVertex(2) = Vector3D<float>(-_widthX/2, _widthY/2,0);
+    (*mesh)[3].getVertex(0) = Vector3D<float>( 0, 0, static_cast<float>(_height));
+    (*mesh)[3].getVertex(1) = Vector3D<float>(static_cast<float>(_widthX/2.0), static_cast<float>(_widthY/2.0), 0);
+    (*mesh)[3].getVertex(2) = Vector3D<float>(static_cast<float>(-_widthX/2.0), static_cast<float>(_widthY/2.0), 0);
 
-    (*mesh)[4].getVertex(0) = Vector3D<float>( 0, 0, _height);
-    (*mesh)[4].getVertex(1) = Vector3D<float>(-_widthX/2, _widthY/2,0);
-    (*mesh)[4].getVertex(2) = Vector3D<float>(-_widthX/2,-_widthY/2,0);
+    (*mesh)[4].getVertex(0) = Vector3D<float>( 0, 0, static_cast<float>(_height));
+    (*mesh)[4].getVertex(1) = Vector3D<float>(static_cast<float>(-_widthX/2.0), static_cast<float>(_widthY/2.0), 0);
+    (*mesh)[4].getVertex(2) = Vector3D<float>(static_cast<float>(-_widthX/2.0), static_cast<float>(-_widthY/2.0), 0);
 
-    (*mesh)[5].getVertex(0) = Vector3D<float>( 0, 0, _height);
-    (*mesh)[5].getVertex(1) = Vector3D<float>(-_widthX/2,-_widthY/2,0);
-    (*mesh)[5].getVertex(2) = Vector3D<float>( _widthX/2,-_widthY/2,0);
+    (*mesh)[5].getVertex(0) = Vector3D<float>( 0, 0, static_cast<float>(_height));
+    (*mesh)[5].getVertex(1) = Vector3D<float>(static_cast<float>(-_widthX/2.0), static_cast<float>(-_widthY/2.0), 0);
+    (*mesh)[5].getVertex(2) = Vector3D<float>(static_cast<float>( _widthX/2.0), static_cast<float>(-_widthY/2.0), 0);
 
     return mesh;
 }
